# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Min Khaung Tea House & Restaurant Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: tableId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/items:
    post:
      operationId: addOrderItem
      tags: [orders]
      summary: Add item to an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddOrderItemBody"
      responses:
        "201":
          description: Added order item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
  /orders/{id}/items/{itemId}:
    patch:
      operationId: updateOrderItem
      tags: [orders]
      summary: Update an order item (status/quantity)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderItemBody"
      responses:
        "200":
          description: Updated order item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderItem"
    delete:
      operationId: removeOrderItem
      tags: [orders]
      summary: Remove an item from order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── KITCHEN ───────────────────────────────────────────────────────────────
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payments
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Process a payment for an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found

  # ─── INVENTORY ─────────────────────────────────────────────────────────────
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff members
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Create staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffMemberBody"
      responses:
        "201":
          description: Created staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    patch:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffMemberBody"
      responses:
        "200":
          description: Updated staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Delete staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── FINANCE ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableNumber:
          type: string
        status:
          type: string
          description: "open, ready_to_pay, paid, cancelled"
        subtotal:
          type: string
        airconFee:
          type: string
        taxAmount:
          type: string
        totalAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tableId
        - tableNumber
        - status
        - subtotal
        - airconFee
        - taxAmount
        - totalAmount
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: string
        customizations:
          type: ["string", "null"]
        kitchenStatus:
          type: string
          description: "new, cooking, ready, served"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - orderId
        - menuItemId
        - menuItemName
        - quantity
        - unitPrice
        - kitchenStatus
        - createdAt
    OrderDetail:
      type: object
      properties:
        id:
          type: integer
        tableId:
          type: integer
        tableNumber:
          type: string
        status:
          type: string
        subtotal:
          type: string
        airconFee:
          type: string
        taxAmount:
          type: string
        totalAmount:
          type: string
        paymentMethod:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        staffId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tableId
        - tableNumber
        - status
        - subtotal
        - airconFee
        - taxAmount
        - totalAmount
        - items
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        tableId:
          type: integer
        notes:
          type: string
        staffId:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - tableId
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        paymentMethod:
          type: string
    AddOrderItemBody:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        customizations:
          type: string
        notes:
          type: string
      required:
        - menuItemId
        - quantity
    UpdateOrderItemBody:
      type: object
      properties:
        quantity:
          type: integer
        kitchenStatus:
          type: string
        customizations:
          type: string
        notes:
          type: string
    OrderItemInput:
      type: object
      properties:
        menuItemId:
          type: integer
        quantity:
          type: integer
        customizations:
          type: string
        notes:
          type: string
      required:
        - menuItemId
        - quantity

    # ─── KITCHEN SCHEMAS ──────────────────────────────────────────────────
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        tableNumber:
          type: string
        amount:
          type: string
        paymentMethod:
          type: string
          description: "cash, wave_pay, kbz_pay, aya_pay, cb_pay"
        status:
          type: string
          description: "pending, completed, refunded"
        cashierId:
          type: ["integer", "null"]
        receiptNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - orderId
        - tableNumber
        - amount
        - paymentMethod
        - status
        - receiptNumber
        - createdAt
    CreatePaymentBody:
      type: object
      properties:
        orderId:
          type: integer
        paymentMethod:
          type: string
        cashierId:
          type: integer
      required:
        - orderId
        - paymentMethod

    # ─── INVENTORY SCHEMAS ────────────────────────────────────────────────
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
          description: "admin, cashier, waitress, kitchen"
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        pin:
          type: ["string", "null"]
        active:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - role
        - active
        - createdAt
        - updatedAt
    CreateStaffMemberBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        pin:
          type: string
      required:
        - name
        - role
    UpdateStaffMemberBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        phone:
          type: string
        email:
          type: string
        pin:
          type: string
        active:
          type: string

    # ─── FINANCE SCHEMAS ──────────────────────────────────────────────────
...