# lib/api-zod/src/generated/types/addOrderItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface AddOrderItemBody {
  menuItemId: number;
  quantity: number;
  customizations?: string;
  notes?: string;
}