# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  tableId: number;
  notes?: string;
  staffId?: number;
  items?: OrderItemInput[];
}