# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Min Khaung Tea House & Restaurant Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  tableId: number;
  tableNumber: string;
  /** open, ready_to_pay, paid, cancelled */
  status: string;
  subtotal: string;
  airconFee: string;
  taxAmount: string;
  totalAmount: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  staffId?: number | null;
  createdAt: string;
  updatedAt: string;
}